Add file
Create a poker.html page, add cards.png and frame.png images from the task folder. It is necessary to cut out the necessary fragments from the frame.png image and make up the rubber background. The size of the frame should depend on the amount of text in it and on the size of the browser window (i.e., the frame stretches to the full width of the browser window, and has such a height to accommodate the text it contains).
Insert title and text. Add a box with a green border and 20px padding. In this container, create 5 blocks (<div>) 190px high and 140px wide and style them display: inline-block. Add a cards.png sprite for these blocks as a background image, but specify such background-position values ​​to get the desired combination of cards.
! Blue is the background color for the body tag, not the border property.