Add file
Create a copy of the poker.html page and name it poker-2.html. Create a new paired <style> tag in the <head> tag and transfer all the styles there using classes, tag selectors and context selectors for elements previously used in the layout. Those. visually, the result in the browser should not change, but the html-code should become clean, without unnecessary attributes, including style.