add file
Create a new page table-test.html in the project and lay out a table in it to get the result as in the picture (see Fig. 1). The table is stretched to fit the entire browser screen in width. The text may differ. The colors used are # 888888, #ffffff and blue, padding for the cells is 5px.