Add files via upload
Create a page chess.html and, using one table, typeset an example of a chessboard (the result is in the picture below). For chess we use the chess.png sprite from the tasks folder. In styles, a call to a sprite should be one single request for all kinds of chess. Think over the layout so that by changing only the classes in the layout, without touching the styles, you can build any chess combination.
Used colors and parameters:
1) background of the document (body) = cyan;
2) the table and title are horizontally centered;
3) the cell size for chess is 48x48 pixels, the background for the "black" field is rgba (0,0,0,0.5);