add file
Create news-section.html page and layout as close as possible to the layouts below. Take arbitrary text and pictures, the size of pictures is no more than 60px by 60px. The maximum block width is 960px, the minimum width is 320px, and is horizontally centered. Minimum width of one news item = 300px