Add files
We continue to work in the Task5 folder.
In it, we create pages first-layout-table.html and first-layout-div.html. It is necessary to lay out the football.png layout (search in the tasks folder) using tables and blocks, respectively.
The layout should be fluid - when changing the width of the browser window, the main column should stretch; the width of the menu column should be fixed and 250px wide. Hyperlinks can lead anywhere. The texts can be arbitrary, but sufficient to allow the layout to be appreciated on a wide screen.
There is no need to stretch the layout to the height of the viewport.