Add file
Place the following text and from the images folder of the image biohazard.png, man-and-woman.png, no-smoking.png, yin-yang.png. Make up the project according to the layout (see at the bottom of the page). Before images - multi-line text for the entire width of the browser window; the images are horizontally centered and adjacent to the text at the bottom, i.e. when the browser window is narrowed, the text height is automatically increased and the images are displaced downward while remaining horizontally centered. Images should be positioned using absolute positioning (position: absolute), image coordinates should be in pixels.
It is also necessary to add a block 200px wide with the text “Your ad could have been here !!!” and position it on the right by 20px and on top by 10% so that when the page scrolls, the block is always in the given place.
Finally, add an “anchor” at the bottom of the page (using the position property) to the first paragraph using a link for which the to-top.png image will be used as a background. The anchor should be located 20px from the bottom border of the viewport'a and centered horizontally.