Add file
Copy the html5-logo.html file to the same Task5 folder.
Without changing the layout and using the original shapes and images (we are looking in the images folder), we achieve the layout as in the layout below. The “result” block must be horizontally centered and have a top margin of 100 pixels. To position the elements (except for the “result” block) on the page, use the “position: absolute;” positioning scheme Do not change the layout and classes, add styles only