Add file
Create a blank-menu.html page and layout as close to the layout as possible. For dropdown menus use position: absolute. Find similar icons, make a sprite out of them and add them to the menu via: before and background-position. Portfolio link view, Submenu 2, Menu level 3 and Submenu 6 - this is what the menu link looks like when you hover over it with the mouse.