Add file
Create adapt-page.html page.
It is necessary to make a layout as close as possible to the layouts (layout1 and layout2, or look for layouts in the tasks folder). You can use arbitrary texts and images. The maximum content width is 980px and is horizontally centered, but the background color of the header and footer must stretch 100% of the screen width. Blocks with news previews should be right and left aligned with the content. Pay attention to the text in the footer, namely its position.
With a screen width of 768px and below, the site should transition to the second layout view. The minimum site width is 320px. When stretching the browser to any extensions larger than 320px, horizontal scrolling should not appear (at all!). Use "overflow: hidden;" prohibited