Add file
Make up the page with the authorization form as in the layout below (Fig1 and Fig2). For the background of the body, use a picture imitating cardboard paper or a board. There should be no more images in the project. The round arrow button should have styles for the: focus and: active states.