Add file
Create a form-styled.html page and copy the html markup and styles from the previous task of this project (form) and change / add styles so as to make the layout as close as possible to the layout of this task. The email field in the layout shows the: focus state for the input field. Choose icons and fonts as close as possible to the layout.
Also, when adding a modifier class to the form (form tag), the fields need to change their position (see Figure 2)