Add file
Create folder Task9.
Add the video-gallery.html page to it
On this page, place a short text description and 4-6 videos from the youtube video service. Choose the topic of the videos yourself, but carefully. Make it so that each video can be viewed both on the page layout, and be able to open the youtube.com video service page with this video in a separate browser tab. Don't forget to make your video responsive!