Add file
We work in the Task9 folder.
Create a page audio-tag.html On this page, place a short text description of a band and make a list of 3 to 5 songs. Choose the group yourself, but carefully. To make it possible to listen to and download each audio recording separately, or to be able to download the entire album (you can use the link to the archive). Below is an approximate layout, it is not necessary to do exactly the same layout, it is better to do something more original and beautiful =)).