Add file
In Task11 folder create knowledges.html page
Make up the elements as close to the layout as possible (see Fig. 1). Then add animation when opening / refreshing the page as shown in the knowledges_progress_bar.gif file (see the tasks folder). Those. animation occurs with a slight delay only when the page is opened or refreshed, not looped.