Add file
We work in the Task12 folder.
We take the grid-gallery.html file as a basis (you can create your own file with your own pictures or content). It is necessary to get as close as possible to the logic of display and behavior from the video file grid-gallery.mp4.
Layout should be fluid and responsive, images should have different sizes and proportions (vertical, horizontal) and maintain these proportions when compressed.